rename to secretprov and improve plugin design
- Rename plugin from keyflux to secretprov across all files (Cargo.toml,
  Makefile, source, docs); update all type names accordingly
- Stop storing private keys after TEE delivery: KeyFluxEntry now holds only
  public material (CA key retained for TLS client cert signing on demand)
- Remove one-set-only guard, has_entry/store_entry helpers, and refresh=true
  param; GET /credentials now regenerates unconditionally
- Switch std::sync::RwLock to tokio::sync::RwLock throughout
- Fix generate_ca_cert missing set_version(2) (X.509 v3 required by RFC 5280)
- Fix P256 branch to respect server_cert_config_store instead of hardcoding CN
- Remove unused constants DEFAULT_SERVER_VALIDITY_DAYS, DEFAULT_CLIENT_VALIDITY_DAYS,
  CA_DEFAULT_COMMON_NAME
- Update docs: remove pod-centric language, fix config field names
  (spec_required, supported_types), remove Security Considerations and
  Troubleshooting sections, replace kbs-client VM examples with REST API form
